# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marlin Frozen Fruits ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchase]
      summary: List all purchases (materials and raw materials)
      responses:
        "200":
          description: Purchases list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchase]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchase]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # ── PRODUCTION ────────────────────────────────────────────────────────────
  /productions:
    get:
      operationId: listProductions
      tags: [production]
      summary: List production batches
      responses:
        "200":
          description: Productions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Production"
    post:
      operationId: createProduction
      tags: [production]
      summary: Create production batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Production"
  /productions/{id}:
    get:
      operationId: getProduction
      tags: [production]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Production batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Production"

  # ── STOCK ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PurchaseLineItem:
      type: object
      required: [materialType, materialId, quantity, unitCost]
      properties:
        materialType:
          type: string
          enum: [material, raw_material, item]
        materialId:
          type: integer
        quantity:
          type: number
        unitCost:
          type: number
        batchNumber:
          type: string
        mfgDate:
          type: string
        expiryDate:
          type: string
    Purchase:
      type: object
      required: [id, vendorId, purchaseDate, lineItems, totalAmount, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        purchaseDate:
          type: string
        invoiceNumber:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseLineItem"
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseInput:
      type: object
      required: [vendorId, purchaseDate, lineItems]
      properties:
        vendorId:
          type: integer
        purchaseDate:
          type: string
        invoiceNumber:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseLineItem"
        notes:
          type: string

    # PRODUCTION
    ProductionLineItem:
      type: object
      properties:
        materialType:
          type: string
          enum: [material, raw_material]
        materialId:
          type: integer
        usedQuantity:
          type: number
    Production:
      type: object
      required: [id, itemId, producedQuantity, productionDate, createdAt]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemName:
          type: string
        producedQuantity:
          type: number
        productionDate:
          type: string
        materialUsed:
          type: array
          items:
            $ref: "#/components/schemas/ProductionLineItem"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductionInput:
      type: object
      required: [itemId, producedQuantity, productionDate, materialUsed]
      properties:
        itemId:
          type: integer
        producedQuantity:
          type: number
        productionDate:
          type: string
        materialUsed:
          type: array
          items:
            $ref: "#/components/schemas/ProductionLineItem"
        notes:
          type: string

    # STOCK
...