# lib/api-zod/src/generated/types/purchaseLineItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Marlin Frozen Fruits ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseLineItemMaterialType } from './purchaseLineItemMaterialType';

export interface PurchaseLineItem {
  materialType: PurchaseLineItemMaterialType;
  materialId: number;
  quantity: number;
  unitCost: number;
  batchNumber?: string;
  mfgDate?: string;
  expiryDate?: string;
}